+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ## Added
 
+- Add can_sign_contract ability to organization
 - Add mermaid graph for Order workflow
 - Add a route to reorder target_courses
 - Add target_course route and list view for products